+++ com.templlo.service.program/src/main/java/com/templlo/service/program/global/aop/distributed_lock/DistributedLockAop.java
@@ -1,6 +1,5 @@
 package com.templlo.service.program.global.aop.distributed_lock;
 
-import com.templlo.service.program.validation.CustomSpELParser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
